inference-server/pyproject.toml
[project]
name = "inference-server"
version = "0.1.0"
description = "Succulent genus classification inference server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]
